move server part from Sagas to server.js, move all actions with Storage to storageFuction(like baseQueries on Sever), organise errors

+++ frontend/src/utils/storageFunctions.js
@@ -1,8 +1,50 @@
 import { v4 as uuidv4 } from 'uuid';
 import moment from 'moment';
+import _ from 'lodash';
+
+export const registerUser = (data) => {
+    const users = getUsersFromStorage();
+    const user = _.find(users, { email: data.email });
+    if (!_.isUndefined(user)) throw new Error('EMAIL_BUSY');
+    const userId = uuidv4();
+    users.push({ ...data, userId });
+    window.localStorage.setItem('users', JSON.stringify(users));
+    return userId;
+};
+
+export const getUserByEmail = (email) => {
+    const users = getUsersFromStorage();
+    const user = _.find(users, { email: email });
+    if (_.isUndefined(user)) throw new Error('NOT_EXIST');
+    return user;
+};
+
+export const getUserById = (userId) => {
+    const users = getUsersFromStorage();
+    const user = _.find(users, { userId: userId });
+    if (_.isUndefined(user)) throw new Error('Server error');
+    return user;
+};
+
+export const updateUserToken = (userId, token) => {
+    const users = getUsersFromStorage();
+    const user = _.find(users, { userId: userId });
+    user.token = token;
+    window.localStorage.setItem('users', JSON.stringify(users));
+};
 
 export const getBoardsFromStorage = () => {
     return JSON.parse(window.localStorage.getItem('boards') || '[]');
+};
+
+export const saveBoardToStorage = (board) => {
+    const boards = JSON.parse(window.localStorage.getItem('boards') || '[]');
+    const boardWithNewName = _.find(boards, { name: board.name });
+    if (!_.isUndefined(boardWithNewName)) throw new Error('Name exists');
+    const newBoardId = uuidv4();
+    boards.push({ ...board, id: newBoardId });
+    window.localStorage.setItem('boards', JSON.stringify(boards));
+    return newBoardId;
 };
 
 export const getListsFromStorageByBoard = (boardId) => {
@@ -11,8 +53,12 @@ export const getListsFromStorageByBoard = (boardId) => {
 
 export const saveListToStorage = (newList) => {
     const lists = JSON.parse(window.localStorage.getItem('lists') || '[]');
-    lists.push(newList);
+    const sameNameList = _.find(lists, { name: newList.name, boardId: newList.boardId });
+    if (!_.isUndefined(sameNameList)) throw new Error('LIST_EXIST');
+    const newListId = uuidv4();
+    lists.push({ ...newList, id: newListId });
     window.localStorage.setItem('lists', JSON.stringify(lists));
+    return newListId;
 };
 
 export const getListsFromStorage = () => {
@@ -21,13 +67,14 @@ export const getListsFromStorage = () => {
 
 export const saveCardToStorage = (newCard) => {
     const cards = JSON.parse(window.localStorage.getItem('cards') || '[]');
-    let lastIndex = 0;
-    cards.forEach((card) => {
-        if (card.listId === newCard.listId && card.index >= lastIndex) lastIndex = card.index + 1;
-    });
-    cards.push({ ...newCard, index: lastIndex });
+    const sameNameCard = _.find(cards, { name: newCard.name, boardId: newCard.boardId });
+    if (!_.isUndefined(sameNameCard)) throw new Error('CARD_EXIST');
+    const cardsInList = cards.filter((card) => card.listId === newCard.listId);
+    const newCardIndex = cardsInList.length;
+    const newCardId = uuidv4();
+    cards.push({ ...newCard, index: newCardIndex, id: newCardId });
     window.localStorage.setItem('cards', JSON.stringify(cards));
-    return lastIndex;
+    return { index: newCardIndex, id: newCardId };
 };
 
 export const getCardsFromStorageByBoard = (boardId) => {
@@ -62,28 +109,60 @@ export const getActivityFromStorageByBoard = (boardId) => {
     return boardActivities;
 };
 
-export const deleteCardsByList = (listId) => {
-    const cards = getCardsFromStorage();
-    const newCards = cards.filter((card) => !listId.includes(card.listId));
-    window.localStorage.setItem('cards', JSON.stringify(newCards));
-};
-
 export const deleteListByBoard = (boardId) => {
     const lists = getListsFromStorage();
-    const deletedLists = [];
-    const newLists = lists.filter((list) => {
-        if (list.boardId === boardId) {
-            deletedLists.push(list.id);
-            return false;
-        } else return true;
-    });
-    deleteCardsByList(deletedLists);
-    window.localStorage.setItem('lists', JSON.stringify(newLists));
+    const listsDevide = _.partition(lists, ['boardId', boardId]);
+    const deletedListsId = listsDevide[0].map((list) => list.id);
+    deleteCardsByList(deletedListsId);
+    window.localStorage.setItem('lists', JSON.stringify(listsDevide[1]));
     const activity = JSON.parse(window.localStorage.getItem('activities') || '[]');
     const newActivity = activity.filter((activity) => activity.boardId !== boardId);
     window.localStorage.setItem('activities', JSON.stringify(newActivity));
 };
 
+export const deleteCardsByList = (listId) => {
+    console.log(listId);
+    const cards = getCardsFromStorage();
+    const newCards = cards.filter((card) => !listId.includes(card.listId));
+    window.localStorage.setItem('cards', JSON.stringify(newCards));
+};
+
 export const getUsersFromStorage = () => {
     return JSON.parse(window.localStorage.getItem('users') || '[]');
+};
+
+export const renameBoard = (data) => {
+    const boards = getBoardsFromStorage();
+    const board = _.find(boards, { id: data.id, user: data.author });
+    if (_.isUndefined(board)) throw new Error('Yoy cant rename this board');
+    board.name = data.name;
+    window.localStorage.setItem('boards', JSON.stringify(boards));
+};
+
+export const deleteBoard = (data) => {
+    const boards = getBoardsFromStorage();
+    const newBoards = boards.filter((element) => element.id !== data.id || element.user !== data.author);
+    if (newBoards.length === boards.length) throw new Error('Yoy cant delete this board');
+    window.localStorage.setItem('boards', JSON.stringify(newBoards));
+};
+
+export const renameList = (data) => {
+    const lists = getListsFromStorage();
+    const listWithNewName = _.find(lists, { name: data.name, boardId: data.board });
+    if (!_.isUndefined(listWithNewName)) throw new Error('Name exists');
+    const listToRename = _.find(lists, { id: data.listId, user: data.user });
+    listToRename.name = data.name;
+    window.localStorage.setItem('lists', JSON.stringify(lists));
+};
+
+export const deleteList = (data) => {
+    const lists = getListsFromStorage();
+    const newLists = lists.filter((element) => element.id !== data.listId);
+    window.localStorage.setItem('lists', JSON.stringify(newLists));
+};
+
+export const deleteCard = (data) => {
+    const cards = getCardsFromStorage();
+    const newCards = cards.filter((element) => element.id !== data.card);
+    window.localStorage.setItem('cards', JSON.stringify(newCards));
 };
